pyproject.toml
[tool.poetry]
name = "final_engineering_project"
version = "0.1.0"
description = ""
authors = ["Yuval Saraf <unimonkiez@gmail.com>"]

[tool.poetry.scripts]
start = "final_engineering_project.index:start"
download = "final_engineering_project.index:download"

[tool.poetry.dependencies]
python = "^3.7"
torch = "1.7.1"
python-dotenv = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.3"
black = "^20.8b1"
mypy = "^0.782"
wemake-python-styleguide = "^0.14.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
